requirements.txt
pytorch
pprint
json
tensorflow
tensorboardX
tensorboard
